requirements.txt
Django==1.8
Pillow>=2.0.0
django-markdown==0.2.1
django-markwhat==1.4
psycopg2==2.5
pycrypto==2.6
simplejson==3.3.0
django-tinymce>=2.2.0
